Renamed config and discovery packages to internal in the HarmonyHub binding. See openhab/static-code-analysis#100
