+++ addons/binding/org.openhab.binding.harmonyhub/OSGI-INF/HarmonyHubDiscovery.xml
@@ -10,7 +10,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.openhab.binding.harmonyhub.discovery">
-	<implementation class="org.openhab.binding.harmonyhub.discovery.HarmonyHubDiscoveryService"/>
+	<implementation class="org.openhab.binding.harmonyhub.internal.discovery.HarmonyHubDiscoveryService"/>
 	<service>
 		<provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
 	</service>
