+++ addons/binding/org.openhab.binding.harmonyhub/src/main/java/org/openhab/binding/harmonyhub/internal/HarmonyHubHandlerFactory.java
@@ -31,9 +31,9 @@ import org.eclipse.smarthome.core.thing.type.ChannelType;
 import org.eclipse.smarthome.core.thing.type.ChannelTypeProvider;
 import org.eclipse.smarthome.core.thing.type.ChannelTypeUID;
 import org.openhab.binding.harmonyhub.HarmonyHubBindingConstants;
-import org.openhab.binding.harmonyhub.discovery.HarmonyDeviceDiscoveryService;
 import org.openhab.binding.harmonyhub.handler.HarmonyDeviceHandler;
 import org.openhab.binding.harmonyhub.handler.HarmonyHubHandler;
+import org.openhab.binding.harmonyhub.internal.discovery.HarmonyDeviceDiscoveryService;
 import org.osgi.framework.ServiceRegistration;
 
 import com.google.common.collect.Sets;
